requirements.txt
pyenv virtualenvasttokens==3.0.0
decorator==5.2.1
exceptiongroup==1.2.2
executing==2.2.0
greenlet==3.1.1
ipython==8.34.0
ipython-genutils==0.2.0
ipython-sql==0.5.0
jedi==0.19.2
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
prettytable==3.16.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
six==1.17.0
SQLAlchemy==2.0.40
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.13.1
wcwidth==0.2.13
